requirements.txt
langchain
langchain-community
sentence-transformers
faiss-cpu
pypdf
fastapi
uvicorn
streamlit
python-dotenv
